# crypto/merkle: optimize merkle tree hashing (#6513) (#9446)