--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [#21](https://github.com/osmosis-labs/cometbft/pull/21) Move websocket logs to Debug
 * [#22](https://github.com/osmosis-labs/cometbft/pull/22) Fix txSearch pagination performance issues
+* [#25](https://github.com/osmosis-labs/cometbft/pull/25) Optimize merkle tree hashing
 
 ## v0.37.4-v24-osmo-2